Overhaul script to tidy up after egs-parallel runs
Improve the script robustness, in particular by forcing the user to
specify either the -n (--dry-run) option, or the -f (--force) option to
actually remove files, to prevent accidental erasing to some extent.
This script removes files without warnings (when using -f), so use with
caution: run with the -n option first to see what will be deleted.

Add the concatenation and sorting of egs-parallel log messages into the
.egsparallel file for reference. This is useful, because these log
messages may be scattered in different files, for example the .eo files
from pbs. After cleaning, the .egsparallel contains a time-ordered
sequence of messages from egs-parallel and its subscripts.