Add HEN_HOUSE/scripts/bin directory, add to path
Add a bin directory in HEN_HOUSE/scripts and add it to the PATH in the
shell additions scripts. This allows some EGSnrc scripts to be directly
executable the user, without using aliases (which are not inherited by
subshells). The immediate motivation is for the top-level egs-parallel
script, and the egs-parallel-clean script, to become visible on the
path, while the egs-parallel sub-scripts remain in scripts and are not
in the path; normally these should not be invoked directly.