Remove shell additions sourcing in egs-parallel
Do not source the shell additions scripts from within the egs-parallel
sub-scripts, as this is not necessary and not secure. Sourcing was only
needed in the dshtask script to get the path to the EGSnrc executables,
because tasks are launched on the pbs nodes without inheriting the
environment. In this case, simply export the PATH variable via the
pbsdsh qsub script.