Add -x and -v options to egs-parallel-clean
Add -x (--extra) option to clean up egs-parallel log files .egsparallel
and .egsparallel.eo. Although this script always echoes progress to the
terminal, add a -v (--verbose) option to echo the commands that are run
by the script, instead of the more concise messages usually reported.
Internally, add an "action" command to ensure that the log messages
remain up to date with the commands.