Remove dependencies on lock file in egs-parallel
Don't quit the egs-parallel submit scripts if no lock file is found, and
add a -f (--force) option to override existing egsjob or lock file.

The lock file for parallel jobs is managed inside EGSnrc, so the script
should not manage it as well: this creates an obscure correlation
between the code and the script. Moreover, the uniform run control
method does no create a lock file. Previously, the submit script would
quit if there was no lock file, now it merely reports its content. The
top-level egs-parallel script now prevents the run if there is an
.egsjob file OR a .lock file, for the same reason. This can be
overridden with the added --force option.